+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * feat: Add static code diagnostic `prefer-const-border-radius`.
 * Improve static code diagnostic `prefer-extracting-callbacks`, don't trigger on empty function blocks.
 * Improve unused files check, add support for `vm:entry-point` annotation.
+* feat: ignore Flutter builder functions in `prefer-extracting-callbacks` rule.
 
 ## 4.3.3